+++ tests/sim/warehouse/test_warehouse.py
@@ -22,6 +22,26 @@ class TestWarehouse(unittest.TestCase):
         self.assertEqual(warehouse, deepcopy_warehouse)
         self.assertNotEqual(id(warehouse), id(deepcopy_warehouse))
 
+    def test_eq(self):
+        # arrange
+        warehouse = Warehouse()
+
+        # act
+
+        # assert
+        self.assertTrue(warehouse.__eq__(warehouse))
+
+    def test_hash(self):
+        # arrange
+        warehouse_1 = Warehouse()
+        warehouse_2 = Warehouse()
+
+        # act
+
+        # assert
+        self.assertEqual(hash(warehouse_1), hash(warehouse_1))
+        self.assertNotEqual(hash(warehouse_1), hash(warehouse_2))
+
     def test_set_pos_y_floor(self):
         # arrange
         warehouse = Warehouse()